# refactor(core): slim reader utils & streamline CI