--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -15,6 +15,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - Optimized CharStream with length caching for faster lexing
+- Deduplicated numeric literal readers via shared utils
+- Inlined CharStream hot paths for speed
 - Added support for `**` and `**=` operators
 
 ### Fixed