--- a/src/lexer/OctalReader.js
+++ b/src/lexer/OctalReader.js
@@ -1,3 +1,5 @@
+import { isOctalDigit } from './NumericLiteralUtils.js';
+
 export function OctalReader(stream, factory) {
   const startPos = stream.getPosition();
   if (stream.current() !== '0') return null;
@@ -6,13 +8,13 @@ export function OctalReader(stream, factory) {
 
   let idx = stream.index + 2;
   const ch = stream.input[idx];
-  if (ch === undefined || ch < '0' || ch > '7') return null;
+  if (ch === undefined || !isOctalDigit(ch)) return null;
 
   let value = '0' + prefix;
   stream.advance();
   stream.advance();
 
-  while (stream.current() !== null && stream.current() >= '0' && stream.current() <= '7') {
+  while (stream.current() !== null && isOctalDigit(stream.current())) {
     value += stream.current();
     stream.advance();
   }