## Summary
- centralize digit checks in `NumericLiteralUtils`
- inline hot paths in `CharStream`
- keep changelog updated

## Benchmark
```
File | Before (MB/s) | After (MB/s) | % of Before
--- | --- | --- | ---
lexer_engine.js | 1.96 | 2.19 | 112%
template_string_reader.js | 3.11 | 2.14 | 69%
```

Public API is unchanged.


------
https://chatgpt.com/codex/tasks/task_e_68542bc03af88331bd6333d1fc22bff0